refactor(ors_control_app): extract lib/ors and lib/cache helpers
Move ORS and cache helpers out of server/index.ts:
- lib/ors.ts: waitForOrsGraphReady, getExpectedProfiles, DEFAULT_PROFILES
- lib/cache.ts: roadPointsCacheKey/Get/Set, formatUptime

server/index.ts now imports these. Lines down 3,545 -> 3,428.

Phase 1 of holistic-control-app-refactor-final, lib extraction step 2.
No behaviour change. tsc -p tsconfig.server.json passes cleanly.
